# docs/swagger.yaml
basePath: /
definitions:
...
  internal_modules_router_interfaces_http.ErrorResponse:
    properties:
      error:
        type: string
      fields:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
    type: object
...
info:
  contact: {}
  description: AI model proxy with async job support. Powered by Gin, Fx, GORM, and
    AWS Lambda.
  title: AI Proxy API
  version: "1.0"
paths:
...
  /router/{id}/v1/embeddings:
    post:
      consumes:
      - application/json
      description: Accepts an OpenAI /v1/embeddings request and routes it through
        the router's embedding model.
      parameters:
      - description: Router ID
        in: path
        name: id
        required: true
        type: string
      - description: OpenAI embeddings request
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_modules_router_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_modules_router_interfaces_http.ErrorResponse'
      security:
      - BearerAuth: []
      summary: OpenAI-compatible embeddings
      tags:
      - hyperstrate
      - routers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"